Changed the State implementation
The entire State implementation has been reverted to the initial one.
Changing the implementation to use a object/memory pool was unfound, as there were no measurements taken to prove it's better
.
The current implementation uses an ABC to create derived states through a static templated method.
States are stored in a unique_ptr.