Add GameContext and ServiceLocator classes
The Engine class now has methods to create GameContexts and make them active.
Each call of the Engines' Loop methods updates the currently active context.
If the context is not set or its Update methods returns false, the game closes.
The Engine has a ServiceLocator
 class to which it registers all of the subsystems.
The locator is passed to each context so they have references to the subsystems.