Add the Box2D project.
The Box2D project wasn't added before because it was a third party library.
Most other libs added to the engine are already compiled, and thus no PDB files were present.
Box2D adds a twist to this, seeing as i had to compile it myself, and thus do have the PDBs.
Since the lib files encode the path to the PDB files in them, this path should be the same across all computers that clone the repo.
Thus, the easiest solution is to just put the Box2D source into the repo and build it once on each machine.