Change the structure of the project files and directories.
Create a Debug and Release directories in the repository root.
After a build, files are copied to the build specific directory.
For now, only Lua dlls are put in there and added to the repo.
No build-generated files should be added to the repo from these folders,
as each machine will treat them as changed files.