Add SpritesheetCache to conform to new asset control.
SpritesheetCache class inherits from AssetStore with Spritesheet as the parameter, and no load argument parameter.
The Engine class owns the only instance of SpritesheetCache (since there is not platform specific stuff in it, that was the logical location).
Clean up GameContext so it doesn't own a copy of ServiceLocator and ResourceLocator but references to them.
The Engine class owns the only instances of Resource and Service Locator classes.