Migration from VS2012 to Eclipse+MinGW, due to bad C++11 support in VS2012.
A lot of the code has been slightly modified in order to conform to the new compiler.
The code currently builds with both Eclipse and VS2012, since no additional code has been added.
All neccessary files have been added to have the Eclipse projects portable to any other machine.
Removed the need for a .def file for building the renderer DLL, as Eclipse has bad support for it.
However, this way (using __declspec(dllimport/dllexport) is much easier.
Also, precompiled headers are not supported in the Eclipse projects yet,
but it's something that will be looked into.

Important note: in order to get a stable build, all projects MUST be built from scratch in your local
repository, since all output artifacts are not stored in Git.