Small changes in the way updates are propagated down to actions.
GameContext now has input, logic and render methods.
Derived contexts should implement onInput and onLogic if they want to do anything specific in those cases.
The default input and logic methods collect the input events and update the logic.
The render method just calls update on the registered render action.
This is done so that multiple updates can happen, but with a single draw after it.
Beggining of code refactor for Timer classes.