Made a short test of the input system
The input system is now complete. There might be some more tweaking about how
the states, actions and ranges are mapped/bound to outbound intents, but the
core of the system works. Since the MoveUp and MoveDown intents are right now
Pong specific, their handlers are just plain functions in the Pong namespace.
Now what remains is to make components for every piece of data i need to have in there...