Move Box2D into the Physics2d Action.
Everything related to Box2D is moved to the Physics Action, in order
to promote isolation of physics code/engine. All Components related to
the concrete physics engine should be defined in their own header file,
in order to more easily swap out physics engines.
The physics Action is also in charge of rendering debug data, so the
b2DebugDraw class has been moved from Gameplay Context into the Physics2d
Action.
MessageSystem has received an update, removing the multimap as a message
container, and making it to be a map with a vector for the message list.
This is made so the messages can retain their insertion order for any
recepient, while the ordering of recepients is not required.