BUILD: Y, RUN: N - Changes to Cache management
The plan is to have a standardized interface for asset loading.
All Caches should have a similair interface, that handles both first load an reload.
Also, dependencies should be autoloaded if out of order loads are encountered.
Also, loading should have a status result (loaded or not), and an optional output value for the ID.