BUILD: Y, RUN: N - Merged all spritesheet info into one file
All info about a spritesheet is now located in one file, and only through it can any data be reloaded.
The spritesheet cache now acts only as a hub for loading the spritesheet data.
The loading is actually split into image data and animation data load.

Issue: The images displayed on screen are very wrong.