RUNS: Changed the way objects are initialized, made generic Manager
Since most of the Managers so far were almost identical, a choice was made
to extract it as a template class. This also meant that loaders had to be
initialized differently, so now the System class is the one holding all of
the classes, initializing them and passing them around for referencing.