RUNS: Rename util, communication between game and window
The window is no longer passed into the game thread, and instead the game
communicated with the window through a communication buffer, the same way
the window communicates with the game (two separate buffers, one for each
direction).
Renamed util to utiliy because Git was annoying with not
renaming the folder from Util to util (change in case).