BUILDS: Started working on Rainbowland code.
The memory stuff is done and in a place where i feel comfortable with it.
It is usable, control over it is direct, there are no side effects that
are not explicitly shown in written code and it's the most simple solution
to the memory distribution that i could have made.
So forward on, i will be making the game itself, with dips into system
land when i realise i need them (like font rendering, audio playing etc).